minor: pull out FormInner into own component
