+++ packages/react-heartwood-components/src/components/Forms/components/FormBuilder/components/index.js
@@ -1,0 +1,1 @@
+export FormInner from './components/FormInner/FormInner'
